move to metric + simplify
